Extract Save Search button into its own component
